pyproject.toml
[project]
name = "api.tzgs.fun"
version = "0.1.0"
description = "APIs used in self developed projects."
authors = [
    {name = "Allen Tao", email = "allen@tkzt.cn"},
]
dependencies = [
    "fastapi[standard]>=0.115.6",
    "loguru>=0.7.2",
    "fastapi-async-sqlalchemy>=0.6.1",
    "pydantic-settings>=2.3.1",
    "bcrypt>=4.1.3",
    "pyjwt>=2.8.0",
    "fastapi-pagination>=0.12.25",
    "captcha>=0.5.0",
    "qrcode>=7.4.2",
    "redis>=5.0.7",
    "slowapi>=0.1.9",
    "greenlet>=3.1.1",
    "sqlmodel>=0.0.22",
    "alembic>=1.14.0",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[dependency-groups]
pg = [
    "asyncpg>=0.29.0",
    "psycopg2>=2.9.9",
]
